Create workshops.md
The initial workshops CFP page. For consistency, suggest the repository owner to create a folder "call-participation" under "ieeevis.org/year/2019/info/" and move this file into that folder, so that the URL would be  <http://ieeevis.org/year/2019/info/call-participation/workshops>, which is consistent with the practice in previous years.