Update call-for-participation.md
I tried to get all the shortpaper information propagated, but I probably broke a link again. Also, the infovis and scivis paper type files seem to not be in the same place as the vast-paper-types.md file. Not sure if this will cause a problem. I use the webpage feature to create the file and then it puts it "somewhere", not sure how to control that. Sorry web team! Thank you for all the help.