Enable the SV entry on the left_sidebar-2019
The SV program is about to start the first event (T-shirt design contest) for this year. We hope to enable the SV entry to get more visibility for this event. 

P.S., I notice that each item on the left-sidebar has an id and a class="_blank". I'm not very sure what they do, but I followed the format. 

-Fumeng Yang