Create best-poster-awards.md
as a response to several requests we propose a web page for poster awards (awardees + process) similar to the best paper awards page. we (poster co-chairs) would appreciate if it can be shown similarly to best paper awards under side bar "Awards and Recognitions". Attaching the information to the existing best paper awards page would also be fine. thx