Update workshops.md
This is the updated schedule for the timings of workshops. 

Note a change from the original 2020-Program in shared in https://docs.google.com/spreadsheets/d/1lrLln81wQdRZnPYGmBh64Vs3GAPVh8A1/edit#gid=1422628425. The change is that VisGuides is now Sun morning and Visualization in Astrophysics is on Mon afternoon. I will email you about this change.